revised module for OpenPedCan v13 data
This commit includes revision of the R script to generate violin plots of total protein distributions with selected proteins highlighted. The script uses the OpenPedCan v13 data, which is included in the pbta_splicing module v7 data download. The one substantial change in the R script is that the HOPE and GBM cohorts are now in separate data files so must be read in separately and joined. The README file has been revised as well.